% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsIO.R
\name{cds_to_h5}
\alias{cds_to_h5}
\title{The monocle is converted to h5 file}
\usage{
cds_to_h5(cds, h5, assay.names = "RNA")
}
\arguments{
\item{h5}{The h5 file in R}

\item{sce}{The monocle object.}

\item{assay.name}{The 'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\description{
monocle object is converted to h5 file.
}
