% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsIO.R
\name{cds_write_h5}
\alias{cds_write_h5}
\title{The monocle is converted to h5 file}
\usage{
cds_write_h5(cds = NULL, file = NULL, assay.name = NULL)
}
\arguments{
\item{file}{The h5 file.}

\item{assay.name}{The 'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}

\item{sce}{The monocle object.}
}
\description{
monocle object is converted to h5 file.
}
