% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsIO.R
\name{h5_to_cds}
\alias{h5_to_cds}
\title{H5 to the monocle3}
\usage{
h5_to_cds(h5, assay.name = "RNA")
}
\arguments{
\item{h5}{The h5 file in R}

\item{assay.name}{'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\description{
H5 file is converted to the monocle3 obejct
}
