% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIO.R
\name{read_h5}
\alias{read_h5}
\title{H5 to scRNAs-seq analysis object}
\usage{
read_h5(target.object, file, assay.name = "RNA")
}
\arguments{
\item{target.object}{Denotes which object to load.Available options are:
\itemize{
  \item "seurat": converted the h5 file to "seurat object".
  \item "singlecellexperiment": converted the h5 file to "singlecellexperiment object".
  \item "monocle": converted the h5 file to "monocle3 object".
}}

\item{file}{The h5 file}

\item{assay.name}{'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\value{
The single cell analysis mainstream software
}
\description{
Read h5 and converted h5 to the scRNA-seq analysis object
}
