% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat_to_h5}
\alias{seurat_to_h5}
\title{Seurat is converted to h5 file}
\usage{
seurat_to_h5(
  seurat = NULL,
  h5 = NULL,
  assay.name = NULL,
  save.graphs = FALSE,
  save.scale.data = FALSE
)
}
\arguments{
\item{seurat}{The seurat object.}

\item{h5}{The h5 file in R.}

\item{assay.name}{The 'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}

\item{save.graphs}{The knn and snn network isn't saved(FALSE by defualt).}

\item{save.scale.data}{The scale data isn't saved(FALSE by defualt).}
}
\description{
Seurat object is converted to h5 file.
}
