% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat_write_h5}
\alias{seurat_write_h5}
\title{The seurat object to h5}
\usage{
seurat_write_h5(
  seurat = NULL,
  file = NULL,
  assay.name = NULL,
  save.graphs = FALSE,
  save.scale.data = FALSE
)
}
\arguments{
\item{seurat}{The seurat object.}

\item{file}{The h5 file.}

\item{assay.name}{The 'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}

\item{save.graphs}{The knn and snn network isn't saved(FALSE by defualt).}

\item{save.scale.data}{The scale data isn't saved(FALSE by defualt).}
}
\description{
The Seurat object is converted to the h5 file.
}
