% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIO.R
\name{write_h5}
\alias{write_h5}
\title{The scRNAs-seq analysis objec to H5}
\usage{
write_h5(
  object.type,
  data,
  file,
  assay.name = "RNA",
  save.graphs = FALSE,
  save.scale.data = FALSE
)
}
\arguments{
\item{object.type}{Denotes which object to save.Available options are:
\itemize{
  \item "seurat": converted the "seurat object" to the h5 file.
  \item "singlecellexperiment": converted the "singlecellexperiment object" to the h5 file.
  \item "monocle": converted the "monocle3 object" to the h5 file.
}}

\item{data}{The scRNA-seq analysis object data.}

\item{file}{The h5 file}

\item{assay.name}{'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\description{
Write h5 and  the scRNA-seq analysis object converted to h5
}
